core: Extract Galilei communication
