+++ src/core/communication.cpp
@@ -37,7 +37,6 @@
 #include "cuda_interface.hpp"
 #include "energy.hpp"
 #include "event.hpp"
-#include "galilei.hpp"
 #include "global.hpp"
 #include "grid.hpp"
 #include "grid_based_algorithms/lb.hpp"
@@ -53,7 +52,6 @@
 #include <boost/range/algorithm/min_element.hpp>
 #include <boost/serialization/array.hpp>
 #include <boost/serialization/string.hpp>
-#include <boost/serialization/utility.hpp>
 #include <utils/mpi/cart_comm.hpp>
 
 using namespace std;
@@ -368,65 +366,6 @@ int mpi_iccp3m_init() {
   return check_runtime_errors(comm_cart);
 }
 #endif
-
-/***** GALILEI TRANSFORM AND ASSOCIATED FUNCTIONS ****/
-void mpi_kill_particle_motion_slave(int rotation) {
-  local_kill_particle_motion(rotation, cell_structure.local_particles());
-  on_particle_change();
-}
-
-REGISTER_CALLBACK(mpi_kill_particle_motion_slave)
-
-void mpi_kill_particle_motion(int rotation) {
-  mpi_call_all(mpi_kill_particle_motion_slave, rotation);
-}
-
-void mpi_kill_particle_forces_slave(int torque) {
-  local_kill_particle_forces(torque, cell_structure.local_particles());
-  on_particle_change();
-}
-
-REGISTER_CALLBACK(mpi_kill_particle_forces_slave)
-
-void mpi_kill_particle_forces(int torque) {
-  mpi_call_all(mpi_kill_particle_forces_slave, torque);
-}
-
-struct pair_sum {
-  template <class T, class U>
-  auto operator()(std::pair<T, U> l, std::pair<T, U> r) const {
-    return std::pair<T, U>{l.first + r.first, l.second + r.second};
-  }
-};
-
-Utils::Vector3d mpi_system_CMS() {
-  auto const data =
-      mpi_call(Communication::Result::reduction, pair_sum{}, local_system_CMS);
-  return data.first / data.second;
-}
-
-REGISTER_CALLBACK_REDUCTION(local_system_CMS_velocity, pair_sum{})
-
-Utils::Vector3d mpi_system_CMS_velocity() {
-  auto const data = mpi_call(Communication::Result::reduction, pair_sum{},
-                             local_system_CMS_velocity);
-  return data.first / data.second;
-}
-
-REGISTER_CALLBACK_REDUCTION(local_system_CMS, pair_sum{})
-
-void mpi_galilei_transform_slave(Utils::Vector3d const &cmsvel) {
-  local_galilei_transform(cmsvel);
-  on_particle_change();
-}
-
-REGISTER_CALLBACK(mpi_galilei_transform_slave)
-
-void mpi_galilei_transform() {
-  auto const cmsvel = mpi_system_CMS_velocity();
-
-  mpi_call_all(mpi_galilei_transform_slave, cmsvel);
-}
 
 /*********************** MAIN LOOP for slaves ****************/
